refactor: move shared Pod helpers to pod.go
Relocate imagePullSecrets (with pullPolicy/nonRootPodSC) out of
database.go into a shared resources file per review feedback.
